snake2/requirements.txt
appdirs==1.4.0
decorator==4.0.11
Django==1.10.5
ipdb==0.10.2
ipython==5.2.2
ipython-genutils==0.1.0
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
Pillow==4.0.0
